fix: extract rate limiter to separate module for Next.js route compat
Next.js disallows non-standard exports from route files. Move rate
limiting logic to src/lib/rate-limit.ts so the route only exports POST.

Entire-Checkpoint: 2c1aa5b8d55c
