+++ src/lib/rate-limit.ts
@@ -1,0 +1,34 @@
+const RATE_LIMIT_MAX = 5;
+const RATE_LIMIT_WINDOW_MS = 15 * 60 * 1000; // 15 minutes
+
+interface RateLimitEntry {
+  timestamps: number[];
+}
+
+const rateLimitMap = new Map<string, RateLimitEntry>();
+
+export function isRateLimited(userId: string): boolean {
+  const now = Date.now();
+  const entry = rateLimitMap.get(userId);
+
+  if (!entry) {
+    rateLimitMap.set(userId, { timestamps: [now] });
+    return false;
+  }
+
+  // Remove expired timestamps
+  entry.timestamps = entry.timestamps.filter(
+    (ts) => now - ts < RATE_LIMIT_WINDOW_MS
+  );
+
+  if (entry.timestamps.length >= RATE_LIMIT_MAX) {
+    return true;
+  }
+
+  entry.timestamps.push(now);
+  return false;
+}
+
+export function resetRateLimit() {
+  rateLimitMap.clear();
+}
